refactor: moved squad to sources file
