+++ packages/shared/src/components/modals/SquadMemberModal.tsx
@@ -3,12 +3,8 @@ import { useInfiniteQuery } from 'react-query';
 import request from 'graphql-request';
 import { graphqlUrl } from '../../lib/config';
 import { ModalProps } from './common/Modal';
-import {
-  Squad,
-  SQUAD_MEMBERS_QUERY,
-  SquadEdgesData,
-} from '../../graphql/squads';
-import { SourceMemberRole } from '../../graphql/sources';
+import { SQUAD_MEMBERS_QUERY, SquadEdgesData } from '../../graphql/squads';
+import { Squad, SourceMemberRole } from '../../graphql/sources';
 import UserListModal from './UserListModal';
 import { checkFetchMore } from '../containers/InfiniteScrolling';
 import { Button, ButtonSize } from '../buttons/Button';
