+++ packages/shared/src/components/squads/SquadHeaderMenu.tsx
@@ -5,8 +5,7 @@ import AuthContext from '../../contexts/AuthContext';
 import EditIcon from '../icons/Edit';
 import TourIcon from '../icons/Tour';
 import ExitIcon from '../icons/Exit';
-import { Squad } from '../../graphql/squads';
-import { SourceMemberRole } from '../../graphql/sources';
+import { Squad, SourceMemberRole } from '../../graphql/sources';
 import TrashIcon from '../icons/Trash';
 import { useLazyModal } from '../../hooks/useLazyModal';
 import { LazyModal } from '../modals/common/types';
